Remove status service
This service add /status data and its not required
