+++ internal/services/connection_service.go
@@ -5,6 +5,11 @@ import (
 	"github.com/pecigonzalo/kafka-canary/internal/client"
 )
 
+type ConnectionService interface {
+	Open()
+	Close()
+}
+
 type connectionService struct{}
 
 func NewConnectionService(canary canary.Config, connectorConfig client.ConnectorConfig) ConnectionService {
